Fix failed block synchronization in x_add_gap_mul_V_sparse.
The problem: due to the while-loop condition, not all warps enter into the block, causing inter-warp divergence. However, as block synchronization is occuring inside the loop, for the diverged warps, the synchronization may happen on the next state - intermingling results from two different states or actions. The only guarantee is that they synchronize on the same location; not that they are the same _iteration_, To solve this, the loop condition was modified to ensure that inter-warp divergence does not happen there.